replace FrameBuffer with "plus" version; misc. enhancements - allow passing in format/base format/type to builder - backwards-compatibility ctors - re-implement FloatFrameBuffer
